% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_palettes.R
\name{croix_all}
\alias{croix_all}
\title{Visualize all croix palettes}
\usage{
croix_all(palType = c("all", "brand", "croix", "mov"), ...)
}
\arguments{
\item{palType}{Which palette type should be plotted? Choices are
\code{all} which plots all palette types, \code{brand} which plots
only brand-type palettes, \code{croix} which plots only croix-type
palettes, or \code{mov} which plots only movie-based palettes. Defaults
to \code{all}.}

\item{...}{Additional parameters to pass to inner functions.}
}
\description{
Plots all croix palettes simultaneously.
}
