% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_data.R
\docType{data}
\name{croix_pal_data}
\alias{croix_pal_data}
\title{Complete list of croix palettes}
\format{An object of class \code{list} of length 141.}
\usage{
croix_pal_data
}
\description{
Use \code{\link{croix_palette}} to construct palettes of
   desired length.
}
\keyword{datasets}
