% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croix.R
\name{croix_palette}
\alias{croix_palette}
\title{The croix palette generator}
\usage{
croix_palette(name, n = NULL, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. To see list of available palettes,
use \code{names(croix_palettes)}, or use \code{croix_all()} to generate
a plot of all color schemes.}

\item{n}{Number of colors desired.}

\item{type}{How do you want the colors displayed? Current options are
\code{discrete} or \code{continuous}. \code{discrete} will generate
specific colors in the desired palette. \code{continous} will generate
a gradient between the specified colors desired.}
}
\value{
A vector of colours.
}
\description{
A collection of color palettes from LaCroix, movies and
   whatever else that's colorful and interesting.
}
\examples{
# Return vector
crPal <- croix_palette("croix_pure")
crPal

# Visualize palette
crPal <- croix_palette("mov_the_shining")
plot(crPal)

# Get summary statistics of palette
crPal <- croix_palette("mov_akira")
summary(crPal)

# Get a gradient of palette
crPal <- croix_palette("croix_coconut", n = 50, type = "continuous")
plot(crPal)

}
\keyword{colors}
