% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croixHouseKeeping.R
\docType{data}
\name{croix_palettes}
\alias{croix_palettes}
\title{Complete list of croix palettes}
\format{An object of class \code{list} of length 45.}
\usage{
croix_palettes
}
\description{
Use \code{\link{croix_palette}} to construct palettes of
   desired length.
}
\keyword{datasets}
