% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_palettes.R
\name{croix_all}
\alias{croix_all}
\title{Visualize all croix palettes}
\usage{
croix_all(...)
}
\arguments{
\item{...}{Additional parameters to pass to inner functions.}
}
\description{
Plots all croix palettes simultaneously.
}
